requirements.txt
Flask==2.2.2
python-Levenshtein==0.23.0
pytest
pytest-flask
sentence-transformers
torch
selenium
